extract passport setup to model
